Add Karma for testing in a browser environment. (#2)
This is important because there are multiple entries to the project: node and browser. For example, the `events` module exists separately for node and browser. Each of these must be tested separately.

Scripts have been added to test against the browser only, or node + browser. Additionally, IE is not normally tested because it produces errors. A script to test browsers (including legacy browsers) has therefore also been included.

To do:
* Ideally both node and browser tests for events would share testing code. This could help to minimize issues later when tests for one are updated (but not the other).