Update Webpack to allow for multiple entries: browser and node. (#3)
This is required due to the existence of certain plugins that will work differently depending on their environment. For example, events in node are fired and listened to in a different way than those in the browser. Two separate plugins therefore exist for events, and each entry (the respective *-index.html file) must include in the relevant plugins.

Tests have also been added to run in different environments depending on the files being tested. By default, all tests are run in node. To write tests that run in the browser, name the files such that they end with `.test.browser.js`. More details about testing with Karma below.

------
Add Karma for testing in a browser environment. (#2)

This is important because there are multiple entries to the project: node and browser. For example, the `events` module exists separately for node and browser. Each of these must be tested separately.

Scripts have been added to test against the browser only, or node + browser. Additionally, IE is not normally tested because it produces errors. A script to test browsers (including legacy browsers) has therefore also been included.

To do:
* Ideally both node and browser tests for events would share testing code. This could help to minimize issues later when tests for one are updated (but not the other).